# Back/docs/swagger.yaml
basePath: /
definitions:
...
  DTO.CreateDevice:
    properties:
      config:
        $ref: '#/definitions/Models.JSONB'
      ip_address:
        type: string
      location:
        maxLength: 200
        minLength: 20
        type: string
      mac_address:
        maxLength: 12
        minLength: 12
        type: string
      name:
        maxLength: 50
        minLength: 5
        type: string
      section:
        type: string
      user:
        $ref: '#/definitions/DTO.RegisterRequest'
    required:
    - location
    - mac_address
    - name
    - user
    type: object
  DTO.DeviceList:
    properties:
      config:
        $ref: '#/definitions/Models.JSONB'
      created_at:
        type: string
      id:
        type: integer
      ip_address:
        type: string
      last_seen:
        type: string
      location:
        type: string
      mac_address:
        type: string
      name:
        type: string
      section:
        type: string
      status:
        type: string
    type: object
  DTO.ErrorResponse:
    properties:
      error:
        example: خطایی رخ داده است
        type: string
    type: object
...
  DTO.MessageResponse:
    properties:
      message:
        example: عملیات با موفقیت انجام شد
        type: string
    type: object
...
  DTO.RegisterRequest:
    properties:
      name:
        maxLength: 20
        minLength: 3
        type: string
      password:
        maxLength: 100
        minLength: 8
        type: string
      role:
        allOf:
        - $ref: '#/definitions/Models.UserRole'
        description: "\U0001F504 validation جدید"
        enum:
        - admin
        - operator
        - user
        - device
      username:
        maxLength: 20
        minLength: 3
        type: string
    required:
    - name
    - password
    - role
    - username
    type: object
...
  DTO.ReplaceVitrineConfigRequest:
    properties:
      items:
        items:
          $ref: '#/definitions/DTO.VitrineItemInput'
        type: array
      size:
        maximum: 20
        minimum: 1
        type: integer
    required:
    - size
    type: object
  DTO.SendDeviceCommand:
    properties:
      payload:
        additionalProperties: true
        type: object
      type:
        type: string
    required:
    - type
    type: object
  DTO.SummaryContent:
    properties:
      created_at:
        type: string
      id:
        type: integer
      main_img:
        type: string
      summary:
        type: string
      title:
        type: string
    type: object
...
  DTO.UpdateDevice:
    properties:
      config:
        $ref: '#/definitions/Models.JSONB'
      ip_address:
        type: string
      location:
        maxLength: 200
        minLength: 20
        type: string
      mac_address:
        maxLength: 12
        minLength: 12
        type: string
      name:
        maxLength: 50
        minLength: 5
        type: string
      section:
        type: string
      status:
        type: string
    type: object
...
  DTO.UploadImageResponse:
    properties:
      url:
        example: /static/uploads/1234567890.jpg
        type: string
    type: object
...
  DTO.VitrinContentList:
    properties:
      items:
        items:
          $ref: '#/definitions/DTO.VitrineSlot'
        type: array
    type: object
  DTO.VitrineConfig:
    properties:
      items:
        items:
          $ref: '#/definitions/DTO.VitrineItemDetail'
        type: array
      size:
        type: integer
    type: object
  DTO.VitrineItemDetail:
    properties:
      content:
        $ref: '#/definitions/DTO.SummaryContent'
      position:
        type: integer
      source:
        type: string
    type: object
  DTO.VitrineItemInput:
    properties:
      content_id:
        type: integer
      position:
        minimum: 1
        type: integer
    required:
    - content_id
    - position
    type: object
  DTO.VitrineSlot:
    properties:
      content:
        $ref: '#/definitions/DTO.SummaryContent'
      is_auto:
        description: true یعنی این اسلات خودکار از آخرین اخبار پر شده
        type: boolean
      position:
        type: integer
      source:
        description: daralouWeb | daralouOperator
        type: string
    type: object
  Models.JSONB:
    additionalProperties: true
    type: object
  Models.UserRole:
    enum:
    - admin
    - operator
    - user
    - device
    type: string
    x-enum-varnames:
    - RoleAdmin
    - RoleOperator
    - RoleUser
    - RoleDevice
info:
  contact: {}
  description: مستندات API پروژه Back (احراز هویت، مدیریت محتوا، مدیریت دستگاه‌ها).
  title: Back API
  version: "1.0"
paths:
...
  /devices:
    get:
      description: لیست دستگاه‌های ثبت‌شده - ادمین همه‌ی دستگاه‌ها را می‌بیند، اپراتور
        فقط دستگاه‌هایی که به او اختصاص داده شده
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/DTO.DeviceList'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/DTO.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/DTO.ErrorResponse'
      security:
      - BearerAuth: []
      summary: لیست دستگاه‌ها
      tags:
      - Devices
    post:
      consumes:
      - application/json
      description: ساخت یک دستگاه به‌همراه کاربر مربوطه (نقش device) - فقط ادمین
      parameters:
      - description: اطلاعات دستگاه
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/DTO.CreateDevice'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/DTO.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/DTO.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/DTO.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/DTO.ErrorResponse'
      security:
      - BearerAuth: []
      summary: ایجاد دستگاه
      tags:
      - Devices
  /devices/{device_id}/vitrine:
    get:
      description: تعداد اسلات‌ها و آیتم‌های پین‌شده‌ی ویترین این دستگاه (برای ادیتور
        پنل ادمین)
      parameters:
      - description: شناسه دستگاه
        in: path
        name: device_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/DTO.VitrineConfig'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/DTO.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/DTO.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/DTO.ErrorResponse'
      security:
      - BearerAuth: []
      summary: دریافت تنظیمات ویترین یک دستگاه
      tags:
      - Vitrine
    put:
      consumes:
      - application/json
      description: جایگزینی کامل تعداد اسلات‌ها و آیتم‌های پین‌شده‌ی ویترین این دستگاه؛
        اسلات‌های پین‌نشده خودکار با آخرین اخبار پر می‌شوند
      parameters:
      - description: شناسه دستگاه
        in: path
        name: device_id
        required: true
        type: integer
      - description: تعداد اسلات‌ها و آیتم‌های پین‌شده
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/DTO.ReplaceVitrineConfigRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/DTO.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/DTO.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/DTO.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/DTO.ErrorResponse'
      security:
      - BearerAuth: []
      summary: تنظیم ویترین یک دستگاه
      tags:
      - Vitrine
  /devices/{device_id}/vitrine/preview:
    get:
      description: همون چیزی که خودِ دستگاه از GET /contents/vitrin می‌بینه، برای
        هر device_id دلخواه (برای پیش‌نمایش در پنل ادمین)
      parameters:
      - description: شناسه دستگاه
        in: path
        name: device_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/DTO.VitrinContentList'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/DTO.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/DTO.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/DTO.ErrorResponse'
      security:
      - BearerAuth: []
      summary: پیش‌نمایش ویترین رندرشده‌ی یک دستگاه
      tags:
      - Vitrine
  /devices/{id}:
    delete:
      description: حذف یک دستگاه بر اساس شناسه - فقط ادمین
      parameters:
      - description: شناسه دستگاه
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/DTO.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/DTO.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/DTO.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/DTO.ErrorResponse'
      security:
      - BearerAuth: []
      summary: حذف دستگاه
      tags:
      - Devices
    put:
      consumes:
      - application/json
      description: ویرایش اطلاعات یک دستگاه - فقط ادمین
      parameters:
      - description: شناسه دستگاه
        in: path
        name: id
        required: true
        type: integer
      - description: فیلدهای قابل تغییر
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/DTO.UpdateDevice'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/DTO.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/DTO.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/DTO.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/DTO.ErrorResponse'
      security:
      - BearerAuth: []
      summary: به‌روزرسانی دستگاه
      tags:
      - Devices
  /devices/{id}/commands:
    post:
      consumes:
      - application/json
      description: از طریق اتصال Socket.IO فعلیِ دستگاه یک دستور فوری برایش می‌فرستد
        (مثلاً باز کردن پنل ادمین بدون رمز) - فقط ادمین
      parameters:
      - description: شناسه دستگاه
        in: path
        name: id
        required: true
        type: integer
      - description: نوع و داده دستور
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/DTO.SendDeviceCommand'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/DTO.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/DTO.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/DTO.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/DTO.ErrorResponse'
      security:
      - BearerAuth: []
      summary: ارسال دستور بلادرنگ به دستگاه
      tags:
      - Devices
  /uploads/images:
    post:
      consumes:
      - multipart/form-data
      description: آپلود یک فایل تصویر (jpg/png/gif/webp، حداکثر ۵ مگابایت) و دریافت
        URL عمومی قابل‌استفاده در main_img/img_list محتوا
      parameters:
      - description: فایل تصویر
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/DTO.UploadImageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/DTO.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/DTO.ErrorResponse'
      security:
      - BearerAuth: []
      summary: آپلود تصویر
      tags:
      - Uploads
securityDefinitions:
  BearerAuth:
    description: توکن دسترسی رو به‌صورت "Bearer {access_token}" وارد کن.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"